refactor: di-v2 모듈의 패키지명을 di로 변경
