+++ app/src/main/java/woowacourse/shopping/ui/MainViewModel.kt
@@ -4,8 +4,8 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.example.di_v2.annotation.Database
-import com.example.di_v2.annotation.Inject
+import com.example.di.annotation.Database
+import com.example.di.annotation.Inject
 import kotlinx.coroutines.launch
 import woowacourse.shopping.model.Product
 import woowacourse.shopping.model.repository.CartRepository
