+++ app/src/main/java/woowacourse/shopping/ui/cart/CartActivity.kt
@@ -8,8 +8,8 @@ import androidx.appcompat.app.AppCompatActivity
 import androidx.core.view.ViewCompat
 import androidx.core.view.WindowInsetsCompat
 import androidx.lifecycle.ViewModelProvider
-import com.example.di_v2.DIContainer
-import com.example.di_v2.ViewModelInjectionFactory
+import com.example.di.DIContainer
+import com.example.di.ViewModelInjectionFactory
 import woowacourse.shopping.R
 import woowacourse.shopping.ShoppingApplication
 import woowacourse.shopping.databinding.ActivityCartBinding
