+++ di-v2/src/main/java/com/example/di/ScopedLifecycleObserver.kt
@@ -1,4 +1,4 @@
-package com.example.di_v2
+package com.example.di
 
 import androidx.lifecycle.DefaultLifecycleObserver
 import androidx.lifecycle.LifecycleOwner
